maxDelay skipped when options.node is type String
Fix a bug where maxDelay was not honored when node was specified using a string instead of a video element.
this.options.node.buffered is undefined causing the delay code not to be run. 
This can be easily worked around by using a video node instead of a String in the constructor

Additionally, remove this.delay++ as this.delay is undefined

Also, I just wanted to say sincerely - Thank you so much for making this project and giving it away. It's just absolutely perfect, I've spent the last several weeks trying to get live streaming to work and within a day of finding this library everything is solved. It does *exactly* what I was looking for *so* cleanly, it's honestly beautiful. <3